\name{XlsxToPdf}
\alias{XlsxToPdf}
\title{XlsxToPdf}
\usage{
XlsxToPdf(x = "MeuArquivo.xlsx", Sheet = "Plan1", Output = "MeuArquivo.pdf")
}
\description{
Essa função foi criada no intuito de facilitar a transformação de arquivos no formato xlsx em pdf.
    É gerado um script vbs que abre o arquivo indicado e salva como pdf.
}
\examples{
XlsxToPdf(x = "MeuArquivo.xlsx", Sheet = "Plan1", Output = "MeuArquivo.pdf")
}
