% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XlsxToPdf.R
\name{XlsxToPdf}
\alias{XlsxToPdf}
\title{Transforma um arquivo xlsx em pdf.}
\usage{
XlsxToPdf(x, Sheet, Output)
}
\arguments{
\item{x}{Nome e caminho onde está localizado o arquivo xlsx.}

\item{Sheet}{Nome da guia no arquivo xlsx que deseja salvar como pdf.}

\item{Output}{Nome e caminho onde será salvo o arquivo pdf}
}
\description{
Transforma um arquivo xlsx em pdf.
}
\examples{
XlsxToPdf("C:/MeuArquivo.xlsx", "Plan1", "C:/MeuArquivo.pdf")
}
